requirements.txt
# psycopg2==2.8.5 uncomment if you prefer to build from source
psycopg2-binary==2.8.6 # comment if you'll build from source

beautifulsoup4==4.9.0
boto3==1.16.17
botocore==1.19.17
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
channels==2.4.0
codeguru_profiler_agent
color-it==2.1.3
cryptocode
Deprecated
Django==3.0.4
django-admin-interface==0.13.0
django-cors-headers==3.7.0
django-extensions==2.2.9
django-middleware-global-request==0.2.0
django-silk==4.1.0
django-storages
djangorestframework==3.12.1
durations==0.3.3
google-auth-oauthlib==0.4.1
google-api-python-client==1.10.0
google-cloud-texttospeech==2.2.0
django-filter==2.4.0
Jinja2==2.11.2
langdetect==1.0.8
nltk
markdown==3.3.2
pyjwt
Pillow==7.2.0
pytest-django==4.1.0
pytest-mock==3.4.0
python-jose[cryptography]
python-keycloak==0.24.0
python-magic==0.4.22
python-telegram-bot==13.7
PyYAML==3.13
requests==2.23.0
requests-oauthlib==1.3.0
sendgrid==6.4.5
spacy
stripe
textdistance==4.2.0
webpreview==1.6.0

wheel==0.34.2
gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
